requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
Werkzeug==3.0.1
Pillow==10.2.0
python-dotenv==1.0.0
email-validator==2.1.0
groq==0.4.2
